inline dashboard caching tests
